move sanity to utils, with eslint import restrictions
